GS-6564
default-timeout configuration is not honored when using distributed or local transaction managers

svn path=/trunk/openspaces/; revision=52643


Former-commit-id: 9236f2ef3ade90fa2bd8acb78783a2338cf0fc5d

+++ src/main/src/org/openspaces/core/transaction/manager/DistributedJiniTxManagerConfigurer.java
@@ -34,9 +34,9 @@ public class DistributedJiniTxManagerConfigurer {
     }
 
     /**
-     * @see DistributedJiniTransactionManager#setDefaultTimeout(Long)
+     * @see DistributedJiniTransactionManager#setDefaultTimeout(int)
      */
-    public DistributedJiniTxManagerConfigurer defaultTimeout(long defaultTimeout) {
+    public DistributedJiniTxManagerConfigurer defaultTimeout(int defaultTimeout) {
         distributedJiniTransactionManager.setDefaultTimeout(defaultTimeout);
         return this;
     }
